[backend_setup_make_migration] - modify the codes

+++ database/migrations/2024_05_03_122917_create_fees_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('fees', function (Blueprint $table) {
             $table->id();
             $table->string('name', 50);
-            $table->unsignedInteger('fee');
+            $table->double('fee', 5, 2);
             $table->timestamps();
         });
     }
